requirements.txt
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0
pygame>=2.5.0
Pillow>=10.0.0
numpy>=1.24.0
